% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_request}
\alias{data_request}
\title{Data extraction from knowstory server}
\usage{
data_request(token = Sys.getenv("token"), file_path, fields)
}
\arguments{
\item{token}{API KEY (can be in global enviroment)}

\item{file_path}{document path. Should be a local path}

\item{fields}{list with parameters to extract. Require name, description and type (value type. Can be a number, string, boolean, date)}
}
\description{
Data extraction from knowstory server
}
\examples{
token <- "XXXXXXXXXXXXXXXXXXXXX"
set_token(token)

fields_to_extract <- list(
  list(
    name = "Placas",
    description = "Placas string",
    type = "String"
  ),
  list(
    name = "Clase",
    description = "Clase string",
    type = "String"
  ),
  list(
    name = "Marca",
    description = "Marca string",
    type = "String"
  ),
  list(
    name = "Year",
description = "Year number",
 type = "Number"
)
)

data_request(file_path = 'document/path/local', fields = fields_to_extract)
}
